backend/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.7
django-cors-headers==4.4.0
django-dynamic-fixture==4.0.1
django-dynamic-fixtures==0.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-django==4.8.0
sqlparse==0.5.1